fix: inline user add, only admin can add

+++ frontend/src/components/UserList.js
@@ -1,14 +1,24 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 
 function UserList() {
   const [users, setUsers] = useState([]);
   const [editingUser, setEditingUser] = useState(null);
   const [editedUser, setEditedUser] = useState({});
   const [departments, setDepartments] = useState([]);
-  const [userRole, setUserRole] = useState(localStorage.getItem("role") || null);
-  const isAdmin = userRole.toLowerCase() === "admin";
+  const navigate = useNavigate()
+  const userRole = localStorage.getItem("role");
+  const [newUser, setNewUser] = useState({
+    firstName: "",
+    lastName: "",
+    role: "",
+    emailAddress: "",
+    departmentName: "",
+  });
+  const [showAddForm, setShowAddForm] = useState(false);
+  
+  
 
   useEffect(() => {
     axios.get("http://localhost:8080/company/users")
@@ -21,18 +31,29 @@ function UserList() {
   }, []);
 
   const deleteUser = (id) => {
-    const confirm = window.confirm("Do you want to delete user?")
-    if(confirm){
-      axios.delete(`http://localhost:8080/company/users/${id}`)
-      .then(() => setUsers(users.filter((user) => user.id !== id)))
-      .catch((error) => console.error(error));
+
+    if(userRole === "Admin" || userRole === "admin"){
+      const confirm = window.confirm("Do you want to delete user?")
+        if(confirm){
+          axios.delete(`http://localhost:8080/company/users/${id}`)
+          .then(() => setUsers(users.filter((user) => user.id !== id)))
+          .catch((error) => console.error(error));
+    }
+    }else{
+      alert("You do not have permission to delete users!")
     }
 
   };
 
   const handleEdit = (user) => {
-    setEditingUser(user.id);
-    setEditedUser({...user});
+
+    if(userRole === "Admin" || userRole === "admin"){
+      setEditingUser(user.id);
+      setEditedUser({...user});
+    }else{
+      alert("You do not have permission to edit users!")
+    }
+
   };
 
   const handleCancel = () => {
@@ -58,21 +79,61 @@ function UserList() {
           console.error(error);
           alert("Failed to update!");
          })
+  };
+
+  const handleAdd = () => {
+    
+    if(!newUser.firstName || !newUser.lastName || !newUser.role || !newUser.emailAddress){
+      alert("Fill out all required fields!")
+      return;
+    }
+
+    const department = departments.find(
+      (dept) => dept.id === parseInt(newUser.departmentName)
+    );
+
+    axios.post("http://localhost:8080/company/users", {
+      ...newUser,
+      depId: {id: newUser.departmentName},
+    })
+    .then((response) => {
+      setUsers((prev) => [...prev, {...response.data, departmentName: department ? department.depName : "N/A",},]);
+      setNewUser({
+        firstName: "",
+        lastName: "",
+        role: "",
+        emailAddress: "",
+        departmentName: "",
+      });
+      setShowAddForm(false);  
+    })
+    .catch((error) => {
+      alert("Failed to add user!")
+    });
+    
+    
   }
 
   const handleInputChange = (e) => {
     const {name, value} = e.target;
-    setEditedUser((prev) => ({
-      ...prev,
-      [name]: value,
-    }))
+    if(editingUser){
+      setEditedUser((prev) => ({
+        ...prev,
+        [name]: value,
+      }))
+    }else{
+      setNewUser((prev) =>({
+        ...prev,
+        [name]: value,
+      }));
+    }
   };
 
 
   return (
     <div>
       <h1>User List</h1>
-      {isAdmin &&<Link to="/add" className="btn btn-primary mb-3">Add User</Link>}
+      
       <table className="table table-bordered">
         <thead>
           <tr>
@@ -90,7 +151,6 @@ function UserList() {
             <tr key={user.id}>
               {editingUser === user.id ? (
                 <>
-                  {/* Inline Edit Form */}
                   <td>{user.id}</td>
                   <td>
                     <input
@@ -158,18 +218,86 @@ function UserList() {
                   <td>{user.emailAddress}</td>
                   <td>{user.departmentName || "N/A"}</td>
                   <td>
-                    {isAdmin && (
-                      <>
-                        <button style={{marginRight: 5}} className="btn btn-success" onClick={() => handleEdit(user)}>Edit</button>
-                        <button className="btn btn-danger" onClick={() => deleteUser(user.id)}>Delete</button>
-                      </>)}
+                      <button style={{marginRight: 5}} className="btn btn-success" onClick={() => handleEdit(user)}>Edit</button>
+                      <button className="btn btn-danger" onClick={() => deleteUser(user.id)}>Delete</button>
                   </td>
                 </>
               )}
             </tr>
           ))}
+          {showAddForm && (
+            <tr>
+              <td>New</td>
+              <td>
+                <input
+                  type="text"
+                  name="firstName"
+                  value={newUser.firstName}
+                  onChange={handleInputChange}
+                  placeholder="First Name"
+                  required
+                />
+              </td>
+              <td>
+                <input
+                  type="text"
+                  name="lastName"
+                  value={newUser.lastName}
+                  onChange={handleInputChange}
+                  placeholder="Last Name"
+                  required
+                />
+              </td>
+              <td>
+                <input
+                  type="text"
+                  name="role"
+                  value={newUser.role}
+                  onChange={handleInputChange}
+                  placeholder="Role"
+                />
+              </td>
+              <td>
+                <input
+                  type="email"
+                  name="emailAddress"
+                  value={newUser.emailAddress}
+                  onChange={handleInputChange}
+                  placeholder="Email"
+                  required
+                />
+              </td>
+              <td>
+              <select
+                name="departmentName"
+                value={newUser.departmentName}
+                onChange={handleInputChange}
+              >
+                <option value="">Select Department</option>
+                {departments.map((dept) => (
+                  <option key={dept.id} value={dept.id}>
+                    {dept.depName}
+                  </option>
+                ))}
+              </select>
+              </td>
+              <td>
+                <button className="btn btn-success" onClick={handleAdd}>
+                  Save
+                </button>
+              </td>
+            </tr>
+          )}
         </tbody>
       </table>
+      {(userRole === "Admin" || userRole === "admin") &&(
+        <button
+        className="btn btn-primary mb-3"
+        onClick={() => setShowAddForm(!showAddForm)}
+        >
+          {showAddForm ? "Cancel" : "Add User"}
+        </button>
+      )}
     </div>
   );
 }
